requirements.txt
neo4j
requests
sentence-transformers
transformers
torch
faiss-cpu
scikit-learn
python-dotenv
click
tqdm
